changed file structure and cleaned testimonial code
