+++ src/App.js
@@ -1,7 +1,7 @@
 import { Routes, Route } from "react-router-dom";
 import Home from "./pages/Home";
 
-import "./App.css";
+import "./App.scss";
 
 function App() {
   return (
