+++ src/components/Card/Card.react.jsx
@@ -1,7 +1,7 @@
 // Import npm packages
 import PropTypes from "prop-types";
 
-import "../css/EventCard.scss";
+import "./Card.scss";
 
 // Import other packages
 import LiveDot from "@mui/icons-material/FiberManualRecord";
@@ -19,7 +19,7 @@ import Chip from "@mui/material/Chip";
 
 import defaultImage from "../../assets/default-card-image.jpg";
 
-EventCard.propTypes = {
+CustomCard.propTypes = {
   //=======================================
   // Component Specific props
   //=======================================
@@ -32,7 +32,7 @@ EventCard.propTypes = {
   }),
 };
 
-EventCard.defaultProps = {
+CustomCard.defaultProps = {
   //=======================================
   // Component Specific props
   //=======================================
@@ -45,7 +45,7 @@ EventCard.defaultProps = {
   },
 };
 
-export default function EventCard(props) {
+export default function CustomCard(props) {
   let status = props.content.status.toLowerCase();
   let statusColor =
     status === "upcoming"
