+++ src/components/Card/Card.scss
@@ -10,8 +10,7 @@
   overflow: hidden;
   text-overflow: ellipsis;
 }
-// For Testimonials
-.line-clamp-testimonail{
+.line-clamp-testimonial {
   display: -webkit-box;
   -webkit-line-clamp: 5;
   -webkit-box-orient: vertical;
