+++ src/components/Events/Events.react.jsx
@@ -1,7 +1,7 @@
 // Import npm packages
 import { Typography } from "@mui/material";
 import PropTypes from "prop-types";
-import EventCard from "./Card/EventCard.react";
+import CustomCard from "../Card/Card.react";
 
 Events.propTypes = {
   //=======================================
@@ -100,7 +100,7 @@ export default function Events(props) {
               }}
               key={index}
             >
-              <EventCard content={items} />
+              <CustomCard content={items} />
             </div>
           );
         })}
