+++ src/components/Navbar/Navbar.react.jsx
@@ -10,7 +10,7 @@ import MenuIcon from "@mui/icons-material/Menu";
 import Toolbar from "@mui/material/Toolbar";
 import Button from "@mui/material/Button";
 import logo from "../../assets/Logo.png";
-import CustomizedMenus from "./CustomizedMenus";
+import CustomizedMenus from "./CustomizedMenus.react";
 import Collapse from "@mui/material/Collapse";
 import ExpandLess from "@mui/icons-material/ExpandLess";
 import ExpandMore from "@mui/icons-material/ExpandMore";
