+++ src/components/Testimonial/Testimonial.scss
@@ -1,0 +1,15 @@
+.swiper-slide {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  background: var(--primary-color-light);
+  margin: 4rem 0;
+  padding: 2rem;
+  border-radius: 4px;
+}
+.swiper-button-next,
+.swiper-button-prev {
+  color: var(--secondary-color);
+  top: 92%;
+}
