--- src/index.css
@@ -1,18 +1,0 @@
-* {
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
-}
-body {
-  height: 100%;
-}
-/* Global Variables */
-
-:root {
-  --primary-color: #29af8a;
-  --secondary-color: #435061;
-  --swiper-navigation-size: 36px !important;
-}
-.swiper-pagination-bullet-active {
-  background-color: #000 !important;
-}
