+++ src/pages/Home.js
@@ -1,7 +1,7 @@
 import React from "react";
-import Events from "../components/Events.react";
-import DrawerAppBar from "../components/Navbar/Navbar";
-import Testimonial from "./Testimonial";
+import Events from "../components/Events/Events.react";
+import DrawerAppBar from "../components/Navbar/Navbar.react";
+import Testimonial from "../components/Testimonial/Testimonial.react";
 
 const Home = () => {
   return (
